+++ src/com/ucm/dasi/catan/resource/provider/ConnectionCostProvider.java
@@ -1,34 +1,34 @@
-package com.ucm.dasi.catan.warehouse.provider;
+package com.ucm.dasi.catan.resource.provider;
 
 import java.util.Map;
 import java.util.TreeMap;
 
 import com.ucm.dasi.catan.board.connection.ConnectionType;
+import com.ucm.dasi.catan.resource.IResourceManager;
 import com.ucm.dasi.catan.resource.ResourceType;
-import com.ucm.dasi.catan.warehouse.IWarehouse;
-import com.ucm.dasi.catan.warehouse.Warehouse;
-import com.ucm.dasi.catan.warehouse.exception.NegativeNumberException;
+import com.ucm.dasi.catan.resource.ResourceManager;
+import com.ucm.dasi.catan.resource.exception.NegativeNumberException;
 
 public class ConnectionCostProvider extends CostProvider<ConnectionType> {
     
-    public ConnectionCostProvider(Map<ConnectionType, IWarehouse> costMap) {
+    public ConnectionCostProvider(Map<ConnectionType, IResourceManager> costMap) {
 	super(costMap);
     }
     
     public static ConnectionCostProvider buildDefaultProvider() {
-	TreeMap<ConnectionType, IWarehouse> costMap = new TreeMap<ConnectionType, IWarehouse>();
+	TreeMap<ConnectionType, IResourceManager> costMap = new TreeMap<ConnectionType, IResourceManager>();
 	costMap.put(ConnectionType.Road, buildRoadCost());
 	return new ConnectionCostProvider(costMap);
     }
 
-    private static IWarehouse buildRoadCost() {
+    private static IResourceManager buildRoadCost() {
 	Map<ResourceType, Integer> costMap = new TreeMap<ResourceType, Integer>();
 	costMap.put(ResourceType.Brick, 1);
 	costMap.put(ResourceType.Lumber, 1);
 	try {
-	    return new Warehouse(costMap);
+	    return new ResourceManager(costMap);
 	} catch (NegativeNumberException e) {
-	    return new Warehouse();
+	    return new ResourceManager();
 	}
     }
 }
