+++ src/com/ucm/dasi/catan/resource/provider/ICostProvider.java
@@ -1,7 +1,7 @@
-package com.ucm.dasi.catan.warehouse.provider;
+package com.ucm.dasi.catan.resource.provider;
 
-import com.ucm.dasi.catan.warehouse.IWarehouse;
+import com.ucm.dasi.catan.resource.IResourceManager;
 
 public interface ICostProvider<TType extends Comparable<TType>> {
-    IWarehouse getCost(TType type);
+    IResourceManager getCost(TType type);
 }
