+++ src/com/ucm/dasi/catan/resource/provider/StructureCostProvider.java
@@ -1,40 +1,40 @@
-package com.ucm.dasi.catan.warehouse.provider;
+package com.ucm.dasi.catan.resource.provider;
 
 import java.util.Map;
 import java.util.TreeMap;
 
 import com.ucm.dasi.catan.board.structure.StructureType;
+import com.ucm.dasi.catan.resource.IResourceManager;
 import com.ucm.dasi.catan.resource.ResourceType;
-import com.ucm.dasi.catan.warehouse.IWarehouse;
-import com.ucm.dasi.catan.warehouse.Warehouse;
-import com.ucm.dasi.catan.warehouse.exception.NegativeNumberException;
+import com.ucm.dasi.catan.resource.ResourceManager;
+import com.ucm.dasi.catan.resource.exception.NegativeNumberException;
 
 public class StructureCostProvider extends CostProvider<StructureType> {
 
-    public StructureCostProvider(Map<StructureType, IWarehouse> costMap) {
+    public StructureCostProvider(Map<StructureType, IResourceManager> costMap) {
 	super(costMap);
     }
 
     public static StructureCostProvider buildDefaultProvider() {
-	TreeMap<StructureType, IWarehouse> costMap = new TreeMap<StructureType, IWarehouse>();
+	TreeMap<StructureType, IResourceManager> costMap = new TreeMap<StructureType, IResourceManager>();
 	costMap.put(StructureType.City, buildCityCost());
 	costMap.put(StructureType.Settlement, buildSettlementCost());
 	return new StructureCostProvider(costMap);
     }
 
-    private static IWarehouse buildCityCost() {
+    private static IResourceManager buildCityCost() {
 	Map<ResourceType, Integer> costMap = new TreeMap<ResourceType, Integer>();
 	costMap.put(ResourceType.Grain, 2);
 	costMap.put(ResourceType.Ore, 3);
 
 	try {
-	    return new Warehouse(costMap);
+	    return new ResourceManager(costMap);
 	} catch (NegativeNumberException e) {
-	    return new Warehouse();
+	    return new ResourceManager();
 	}
     }
 
-    private static IWarehouse buildSettlementCost() {
+    private static IResourceManager buildSettlementCost() {
 	Map<ResourceType, Integer> costMap = new TreeMap<ResourceType, Integer>();
 	costMap.put(ResourceType.Brick, 1);
 	costMap.put(ResourceType.Grain, 1);
@@ -42,9 +42,9 @@ public class StructureCostProvider extends CostProvider<StructureType> {
 	costMap.put(ResourceType.Wool, 1);
 
 	try {
-	    return new Warehouse(costMap);
+	    return new ResourceManager(costMap);
 	} catch (NegativeNumberException e) {
-	    return new Warehouse();
+	    return new ResourceManager();
 	}
     }
 }
